ajout d'un dialog de preferences pour les langages chercher sur le web. prise en compte de cet preference pour le renommage des fichiers. Cas special de la langue japonaise, si la langue pref est japonnais et que la langue actuelle ne l'est pas, on recherche un nom japonnais sans les KANJI, si on trouve pas, on passe a la langue suivante

+++ src/main/java/renameFiles/metier/web/series/Episode.java
@@ -9,7 +9,7 @@ public class Episode
     private URI url;
     private String name;
     private int season;
-    private int number;
+    private float number;
     private String type;
     private Date airdate;
 
@@ -34,7 +34,7 @@ public class Episode
         return season;
     }
 
-    public int getNumber()
+    public float getNumber()
     {
         return number;
     }
